+++ tensorflow/core/profiler/protobuf/BUILD
@@ -42,6 +42,7 @@ tf_proto_library(
     name = "diagnostics_proto",
     srcs = ["diagnostics.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [
         ":friends",
     ],
@@ -51,6 +52,7 @@ tf_proto_library(
     name = "input_pipeline_proto",
     srcs = ["input_pipeline.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     protodeps = [":diagnostics_proto"],
     visibility = [
         ":friends",
@@ -61,6 +63,7 @@ tf_proto_library(
     name = "overview_page_proto",
     srcs = ["overview_page.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     protodeps = [
         ":diagnostics_proto",
         ":input_pipeline_proto",
@@ -74,6 +77,7 @@ tf_proto_library(
     name = "op_metrics_proto",
     srcs = ["op_metrics.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -81,6 +85,7 @@ tf_proto_library(
     name = "pod_stats_proto",
     srcs = ["pod_stats.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     protodeps = [
         ":diagnostics_proto",
     ],
@@ -91,6 +96,7 @@ tf_proto_library(
     name = "pod_viewer_proto",
     srcs = ["pod_viewer.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     protodeps = [
         ":diagnostics_proto",
         ":pod_stats_proto",
@@ -102,6 +108,7 @@ tf_proto_library(
     name = "steps_db_proto",
     srcs = ["steps_db.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     protodeps = [":op_metrics_proto"],
     visibility = [
         ":friends",
@@ -112,6 +119,7 @@ tf_proto_library(
     name = "op_profile_proto",
     srcs = ["op_profile.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -119,6 +127,7 @@ tf_proto_library(
     name = "op_stats_proto",
     srcs = ["op_stats.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     protodeps = [
         ":diagnostics_proto",
         ":kernel_stats_proto",
@@ -135,6 +144,7 @@ tf_proto_library(
     name = "kernel_stats_proto",
     srcs = ["kernel_stats.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -142,6 +152,7 @@ tf_proto_library(
     name = "tf_function_proto",
     srcs = ["tf_function.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -151,6 +162,7 @@ tf_proto_library(
     name = "tf_stats_proto",
     srcs = ["tf_stats.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -158,6 +170,7 @@ tf_proto_library(
     name = "trace_events_proto",
     srcs = ["trace_events.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -165,6 +178,7 @@ tf_proto_library(
     name = "hardware_types_proto",
     srcs = ["hardware_types.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -172,6 +186,7 @@ tf_proto_library(
     name = "tfstreamz_proto",
     srcs = ["tfstreamz.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -179,6 +194,7 @@ tf_proto_library(
     name = "memory_profile_proto",
     srcs = ["memory_profile.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -186,6 +202,7 @@ tf_proto_library(
     name = "tf_data_stats_proto",
     srcs = ["tf_data_stats.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":friends"],
 )
 
@@ -193,6 +210,7 @@ tf_proto_library(
     name = "memory_viewer_preprocess_proto",
     srcs = ["memory_viewer_preprocess.proto"],
     cc_api_version = 2,
+    make_default_target_header_only = True,
     visibility = [":memory_viewer_friends"],
 )
